Move graphql queries into same file
