--- src/leetcode/graphql/queries/userProfile.graphql.ts
@@ -1,85 +1,0 @@
-import { GraphQLContext, GraphQLQuery } from '../../models';
-import { getGraphqlLink } from '../../utils';
-
-export function getUserProfileContext(username: string): GraphQLContext {
-  const graphqlLink: string = getGraphqlLink();
-
-  const userProfileQuery: GraphQLQuery = {
-    operationName: 'getUserProfile',
-    query: `
-      query getUserProfile($username: String!) {
-        allQuestionsCount {
-          difficulty
-          count
-          __typename
-        }
-        matchedUser(username: $username) {
-          username
-          socialAccounts
-          githubUrl
-          contributions {
-            points
-            questionCount
-            testcaseCount
-            __typename
-          }
-          profile {
-            realName
-            websites
-            countryName
-            skillTags
-            company
-            school
-            starRating
-            aboutMe
-            userAvatar
-            reputation
-            ranking
-            __typename
-          }
-          submissionCalendar
-          submitStats {
-            acSubmissionNum {
-              difficulty
-              count
-              submissions
-              __typename
-            }
-            totalSubmissionNum {
-              difficulty
-              count
-              submissions
-              __typename
-            }
-            __typename
-          }
-          badges {
-            id
-            displayName
-            icon
-            creationDate
-            __typename
-          }
-          upcomingBadges {
-            name
-            icon
-            __typename
-          }
-          activeBadge {
-            id
-            __typename
-          }    __typename
-        }
-      }
-    `,
-    variables: {
-      username,
-    },
-  };
-
-  return {
-    link: graphqlLink,
-    query: userProfileQuery,
-    headers: {},
-  };
-}
